prunes return edges for conditional calls
A predicated invocation of a procedure which is expected to return, must perform two actions under the same predicate - store the return address and actually do the control flow transfer.
E.g., assume the `bleq foo` instruction
```
if (ZF) {
    LR := 4;
    call foo
}
```

That will have the following IR
```
when ZF goto @take_call
goto @next

take_call:
LR :=   4
call @foo return @next
```
next:
When we see an interprocedural control flow which is predicated, then we can safely assume that there weren't any data effects under the same predicate, therefore this call is not expected to return.
Of course, it is possible to craft IR which will violate this property, e.g.,
``
LR := 4
when ZF call @foo return @next
call @bar return @next
next:
```
which corresponds to the following pseudocode `if (a = b) foo() else bar()`
But such representation could be generated only after a global optimization pass or if an architecture actually has an instruction that dispatches several calls based on a condition.
The global optimization is not possible in the context of the sema-lifter as it is applied on a per-instruction basis, so it always operates with the code produced for the given instruction.