represents ARM conditionals with ite-expressions
So far we were using it/then/else statements to represent conditional
movements in ARM, which were, in turn, translated into 3 basic blocks.

This patch will translate them into ite expressions wich leads to a
much more compact representation, especially in optimized ARM code,
which uses a lot of conditional moves.

In the implementation we predicate only moves to physical registers,
and leave temporary variables as it is, since we expect that the
temporary variable won't escape the scope of the predicated
instruction.