fixes a missing block start which led to duplicating code (#989)
The disassembler was occasionally missing a start of a basic block
which was a destination of some jump due to an improper (and as always
premature) optimization. Now this is fixed.

While searching I've found yet another (smaller) bug in which we were
cancelling the tasks and afterwards still adding it to the work
list. It didn't ever manifest, I believe.