tweaks the print plugin
The print plugin is probably the oldest plugin in BAP, and it was
really crying out for a little bit of attention and love. Not that I
was able to give it what it wants, but here are some tweaks that
hopefully will make our life easier (and nicer). Here is the list of
new features

1) `-dcfg` - will now print graphs of all subroutines as subgraphs
   enclosed in partitions, as well as print the interprocedural
   edges. This will also let `xdot` and other tools that didnt'
   understand several digraphs in row to work with bap directly, e.g.,
   `bap ./exe -dcfg | xdot -`

2) `-dgraph` is the new outputting format which is basically the same
   as `-dcfg` but without the IR terms (i.e., only basic blocks, no
   def terms). Also partitioned and with interprocedural edges.

3) tid numbers are properly escaped now (it looks like that dot or
   xdot interprets somehow them even if they are delimited with quotes)

4) new mechanism for filtering the output, instead of old
   `--print-symbol` and `--print-section` (which are still supported) we
   now have a new `--print-matching` filter that accepts
   `<property>:<regex>` format. The `<property>` field could be for
   now one of `section`, `segment`, `symbol`, or `name`. Where
   `symbol` now denotes the name of the symbol take from the file
   symbol table or accompanying debugging information, e.g.,
   dwarf. Not that symbol that we reconstructed during the
   disassembling. And `name` has the same meaning that `symbol` was
   bearing before, i.e., the name of a subroutine. Concerning the
   regular expression, the accepted syntax is PCRE with partial
   matches, e.g., `--print-matching=section:text` will match with
   `.text` and `texting`. Use `\b` to specify word boundaries, e.g.,
   `--print-matching='symbol:\bmain\b' will print only `main` (and not
   `__libc_start_main`). Do not forget to delimit the regex with
   single quotes, to prevent your shell from ringing the bell.

A Bug in Graphviz
=================

A side note on graphviz dot (and xdot). Due to a 20 years old
[bug][1] (supposedly finally fixed in version 2.40) dot is usually unable
to render more or less complex graphs. This bug is triggered randomly
and depends on many variables (like fonts, subroutine names, etc). It
manifests itself as a message (not visible when xdot is used) and
afterwards `dot` (and correspondigly xdot) will just hang up and stale
for infinity
```
Error: trouble in init_rank
	%0 1
	%0 1
	%0 1
        <lots of tid-looking identifiers>
```

I was using version 2.40 (via docker alpine) and was able to render
some complex graphs, but I believe I'm still hitting this issue on
some graphs even in version 2.40.

[1]: https://github.com/ellson/MOTHBALLED-graphviz/issues/1213