abandon data-regions and fiexed few bugs
COFF section header has two fields that denotates size: virtual size and raw data size.
The virtual size is meaningful only for executable files and it's a size
of the section when it's loaded into a memory.
But for the object files this field doesn't have any sense and equal to
zero. That's why we didn't mark memory in Image module correctly and
didn't have any output for -dasm option. So we need to use raw data size instead.

Also fixes wrong addresses for object files.
Our llvm backend provides relative addresses for all the pieces of information (sections, symbols
etc.) and a default base address as well, so we can substitute a default if we need to.
That works fine for executable files. But when we deal with the object files the problem becomes
visible, since we don't actually have any addresses at all and have to use offsets instead.
Previously we were taking offsets as they are, and a first byte of code had an address equal to the
section offset where this byte belonged.
This PR fixes this and use offsets relative to the offset of the first executable section, so the first byte
will have address 0, as it's expected.