starting moving things around
The new Bap_main library will be responsible for unified command line
and configuration file interface, and will make it easier to embed bap
in other frameworks (now BAP will behave more like a library, i.e.,
won't flood in channels, abort or otherwise terminate your
application, etc).

Alghough it is more or less about the command line parsing the main
idea is to make bap as a library independ on the presence of command
line, environment and stuff like that, so that other frameworks that
would like to embed bap in an application which has its own command
line interface (or doesn't have one at all).

Another important change is that the current behavior associated with
the bap frontend (i.e., disassemble a file and apply passes to the
result), will become a plugin by itself. This PR introduces a new
extension point, called `commmand` which is an entry point, sort of
like pluggable `main` procedure.

When uses from a command line, commands are mapped to subcommands,
e.g.,

```
bap disassemble /bin/ls --dump=bir
bap byteweight-train
bap build my.plugin
bap help disassemble
```

Basically, now bap will have the same interface as `docker` or `git`.
(For backward compatibility the `disassemble` command, which
corresponds to the current behavior of `bap` will be made the default,
so that `bap /bin/ls --dump=bir` will be a shortcut to `bap
disassemble /bin/ls`).

When used as a library, the framework could be easily initialized via
`Bap_main.init`, which will take care of loading plugins, configuring
them, etc. It is possible to provide the sources of configuraration,
command line parameters (if desired), environment and even hook up
logging and error handling. While commands could be invoked directly,
it is usually better when a command has two interfaces - the command
line one, and another the programmatical, via an mli. So we are
encouraging extension writers to provide a library interface to their
plugins, when possible.

One of the important consequenses of this change is that it will open
a venue to writing analysis that doesn't depend on the Projec.t data
structure (the one that is very expensive to build). It will also
relax a few dependencies, e.g., now it would be possible to write a
plugin that doesn't depend on Bap.Std. Which will make bap more robust
and make builds faster.