introduces a more fine-granular view on the image memory (#988)
* marks memory either as code or data

The default BAP view on the image memory is a segment view.
In practice, it means that the memory we
disassemble contains both code and data regions, although we need
only former ones.
In `bap.1.x` with recursive disassembler it wasn't a big issue,
since addresses in the data regions unlikely appeared as a destination
of some real instruction.
But the situation is different in `bap.2.0`. A brand-new speculative
disassembler eager to process
as much of the given segment as it can, trying to disassemble
even data regions and spend time to prove that there is not any code there.
The idea is to feed only true code regions to the disassembler to prevent
him from processing the data regions. And all we need to do is to
mark memory regions either as data or code.

There are no tricks for elf files: we already know
what sections contain code and what contain data.
We just needed to spread this information further.

There was a bug in our llvm backend since we ignored
sections with flag `S_ATTR_SOME_INSTRUCTIONS`, although
it didn't affect anything, as we worked with segments.
But now it's crucial to fix it in order not to miss any code.
Also, for the data regions, we need a flag if it's writable or not.
There is no such flag for section, so we take from the segment that
contains this section.

There was a sort of bug too. For some reason sections with data
were marked as executable and I didn't find any reasonable
explanation for that. I.e. sections with flags
`IMAGE_SCN_CNT_INITIALIZED_DATA` and `IMAGE_SCN_CNT_UNINITIALIZED_DATA`
were marked as code sections.
It didn't affect anything as well, since we were processing more data
than we need.
But now we need to fix that since we need strictly to separate
code and data.
Also, Coff files are a bit tough since there is a virtual size for
sections that could be less than the real one or could be even 0 like
in object files. That's the reason we can't process them as easy as
Elf and Macho files.

This PR fixes https://github.com/BinaryAnalysisPlatform/bap/issues/973

* abandon data-regions and fiexed few bugs

COFF section header has two fields that denotates size: virtual size and raw data size.
The virtual size is meaningful only for executable files and it's a size
of the section when it's loaded into a memory.
But for the object files this field doesn't have any sense and equal to
zero. That's why we didn't mark memory in Image module correctly and
didn't have any output for -dasm option. So we need to use raw data size instead.

Also fixes wrong addresses for object files.
Our llvm backend provides relative addresses for all the pieces of information (sections, symbols
etc.) and a default base address as well, so we can substitute a default if we need to.
That works fine for executable files. But when we deal with the object files the problem becomes
visible, since we don't actually have any addresses at all and have to use offsets instead.
Previously we were taking offsets as they are, and a first byte of code had an address equal to the
section offset where this byte belonged.
This PR fixes this and use offsets relative to the offset of the first executable section, so the first byte
will have address 0, as it's expected.

* removed unused code

* set to zero base address for non-exec files

* fixes identation