allows multiples calls to `Extension.declare`
and, correspondigly to `Self().when_ready` which is implemented
through `Extension.declare`.

Previously, consequent calls were erasing registered code, so that
only the latter extension was invoked. We could probably dismiss this
as an error, but it is easier to allow to register several extension
points in the context of the same plugin.