preserves the solution passed to the fixed point solver
The fixed point solver accepts an initial solution, which in turn
could be a solution for the previous attempt. Originally, it was
assumed that the solution is applied to the same graph and that
a solutions could be only refined on the same graph. Therefore, if
applied to a different graph all solution entries that didn't have
a corresponding entry in the new graph were removed from the solution.

It turns out that it is quite useful to apply the same solution to
different graphs even (and especially) when they are not
intersecting.

This PR proposes to keep all entries in the solution passed to the
fixed point computation. It will update those fields that were
changed, leaving those that weren't intact, basically guarantees that
the newly computed solution contains the superset of the old solution.