more precise subroutine packing
The subroutine packing algorithm is responsible for partitioning a set
of basic blocks into a set of disjoint sets each denoting a subroutine,
such that in each set there is a designated block, called entry, which
dominates all other blocks in this set.

In the ideal world, this partitioning is naturally induced by the
dominators tree of a graph. However, a dominator tree is undefined for
a graph that contains unreachable nodes.

Our new subroutine packing algorithm, introduced with BAP 2.0 alpha
release, solved this problem by connecting all unconnected components
to the root node, thus promoting entries of unconnected components to
subroutine entries. This led to a correct but very imprecise result,
e.g., for x86 we have x2 to x10 more partitions than true
subroutines. Mostly due to lots of unreachable small pieces of code,
which were inserted by the compiler after barriers (for the alignment
purposes) - the issue that we underestimated.

Our original intent was to leave it as it is, assuming that later we
can implement analysis that will refine the packing, e.g., by
discovering switch tables or proving that paths are
infeasible. However, it turned out that it is very hard to refine the
partitioning, since this process couldn't be made iterative - i.e.,
the monotonicity requirement allows us only to add new paths in a
graph but not to remove existing. In other words we can't unwind the
damage introduced by connecting an unconnected component to the root
node, so we have to start from scratch every time. To summarize the
new algorithm produces a very conservative but irreparable
disassembly.

The new algorithm is more precise (and therefore less efficient) and
instead of connecting the unconnected components to the root node we
postpone the decision until the later times. Unless there is no
explicit information in the knowledge base, that the node is a
subroutine start, we won't treat it as such. We are still
disassembling everything (using speculative disassembler), but when we
are reifying the knowledge into the IR graph, we only materialize
those subroutines which were present in the knowledge base. Therefore
letting the downstream analysis to discover more subroutines (or pieces
of code) and insert corresponding edges, so that the partitioning
could be monotonically refined.

The algorithm uses the fixed point solution of the domination
relationship, except that the node transfer function is not
transferring the node presence forward if the input is unknown (the
open circuit case). See the algorithm description in the diff section
for more details and formalization.