pushes the raw lifter to the lower level
It now operates on the Ogre level, so it is capable of creating both
images and projects, therefore it could be used for the new objdump
command.

Another nice addition, is that now it is possible to specify file
the offset and length, that enables extracting portions of big
binaries, e.g., the following will extract the `main` function from
our testing echo binary,

```
bap testsuite/bin/x86_64-linux-gnu-echo --loader=raw \
   --raw-{base=0x40076b,offset=0x76b,length=542}
```

The funny part is that our symbolizers will still work correctly and
will be able to provide names.

This features basically make the raw loader a swiss-knife unpacker,
which could be used to run bap on arbitrary files. Besides, the
disassembly of our testing echo is identical with the raw and llvm
lifters :)