prunes Primus leave events on abnormal evaluation
Before this PR we were emitting the `leave-*` events even when a term
didn't evaluate normally (but instead raised an exception). As a result
event handlers were invoked in the context where the term was partially
(if at all evaluate), while they were expecting that the evaluation was
successful. This led to a few undefined variables and other errors.

Even worse, since the `leave-*` events were posted in the exception
handler, it was possible for the event handler to raise an exception on
its own, which was escaping the original exception handler and breaking
havoc and significantly affecting performance.