revamps bap-byteweight tool
The original idea was just to move to the new rails of Bap_main, but I
accidentally did some refactoring and cleanup and a couple of tweaks.

The main functional changes

1) when the compiler is set, then no other entries beyond the selected
compiler are touched or used.

2) files in the training set are deduplicated by their md5 sum, to
prevent repeating entries.

3) it is now possible to choose the loader for images

4) it is now possible to download signatures for the specified
version, e.g., `bap-byteweight fetch --from-version 1.5.0`

The documentation is also (re)written, should be much more clear now
how to use it.