tweeks the Byteweight implementation
We started to use the --no-byteweight option too often recently, an
indicator that we should revisit Byteweight, and give it at least
some renovation.

First of all, we noticed that the new 1.6.0 signatures drastically increased
the false positive rate from more or less acceptable 5%, which we were
getting with the old 1.5.0 signatures, to as high as 50% rate of false
positives, which we are getting now. We will revisit and retrain the
signatures again, and make sure that in the future the signature
updating procedure will be more controlled.

Next, we noticed that during thresholding we're just comparing the
sample probability with the threshold, without using any prior
probability of a function star at all, despite the fact that we have
this information readily available. Not using an available information
is suboptimal, so we decided to devise a new decision procedure, that
will involve it and, at the same time, provide us a more meaningful
threshold value. We chose to use Bayes factors for the hypothesis
testing, which compares the ratio of posterior probabilities of h1 and
h0 hypothesis.

Another update, based on the observation that we had a few very short
prefixes in our signatures, e.g., 0x48. And a classification based on
such a small prefix wouldn't be very reliable. Thus we added a lower
limit to the prefix length. Our experiments showed that ingnoring
prefixes shorter than 8 give the best results, but more experimenting
should be conducted before we can choose an optimal set of parameters
(right now we have min-length 8 and threshold = 10 (strong
evidence)). These parameters give us better results on both new and
old signatures, than the old implementation, with false positive and
false negative rates going below 1%.

To implement this, we had to extend the Trie and Byteweight
interfaces. We added a general purpose iterators to the Trie
interface and different decision interfaces to the Byteweight
interface. To keep backward compatibility, we packed the newly updated
interfaces into the versioned modules.

Additionally, we reimplemented the pretty-printer for signatures (and
tries in general), to make the output much more understandable.

This is the first step in Byteweight Renovation, we will return to it
and (probably) find even better decision procedure (which will take
into account individual probabilities of substrings and other
factors), as well as update the bap-byteweight toolkit with a command
that will automatically compute false positive/false negative
rates. But all this after the release.