runs ALL the unit tests on Travis (#1006)
It turned out that we don't launch all the available unit tests on Travis.

The reason is that we can't launch tests by running `make test` target
on travis, because we use stages and don't even have a code on the stage
with unit tests. Instead, we install tests executables on the
build stage and then manually launch them on the other stages.
And for some reason we install and run only the most important one
but not all of them.

This PR fixes it and run all the tests we currently have.