don't treat `jump next` as fall
TL;DR; Warning: this PR changes the semantics of BAP 1.x branchers, they
shall not reference the fall-through destination in their outputs for
non-control-flow affecting instructions.
Please, consider updating your custom branchers, if you have one.

In BAP 2.0 a set of destinations of an instruction is stored in its
`dests` slot. This set, if known, shouldn't include the fall-through
destination and only control-flow instructions shall have a non-empty
set of destinations. Since previously our branchers were including all
destinations, there was an extra filtering procedure in the
disassembler, which was removing the fall-through destination, in case
if it still present, to prevent issues with old BAP 1.x branchers.

We had to remove this extra step since clang is using the `jmp next`
instruction (where `next` is the address of the first byte following the
instruction) a lot in its unoptimized output. Probably for
instrumentation or as an initial step for jump relaxation. In any case,
`jump next` is possible and is quite often, so we can't just remove the
destination that looks like a fall-through destination, as it will lead
to inconsistencies in the later stage of analysis, in particular, we
will interpret `jump next` as `nop` instruction in the disassembler, but
denote it as a control-flow term during the semantic analysis.

Our default jump-destinations analysis is already correct, i.e.,
it won't emit any false fall-through edges. The only brancher that
should be updated is the IDA brancher. Fortunately, our interface with
IDA services is rich enough so we can distinguish which jump is which on
the BAP side, so there is no need to update the ida-python package here.