First class theories and documentation (#1010)
* an iteration on documentation

* next iteration of documentation

* thrid iteration, knowledge looks finished.

I will go through it the forth time later.
I also hide the Data module from the knowledge is it is currently not
used and not really finished. So it is better to keep it.

* an iteration on documentation

* finishes more or less with bitvector theories.

* started documenting floating point theories.

* finished with the floating point numbers.

* restores tags naming

In BAP 1.x the function Tag.name returned the name that was provided as an argument to the name parameter to the register function, which is not guaranteed to be unique, but is more human-readable.

In BAP 2.0 after values were reimplemented in terms of the Knowledge value, the identifier returned by the Tag.name function became the fully-qualified name with namespace included, which is less readable and, more importantly, breaks the map-terms facility on which we rely a lot on our IDA integration.

This PR restores the old behavior.

* finished bitv, bool, and float grammars

* finished all grammars, now going to parser.

* finishes documenting bap-core-theory

* fixed typeid names

* implements new interface

* brings back old uuid for bap attrs

* introduces the Knowledge.Name module

This module handles fully qualified names which are ubiquitous in
bap.

* fixes a bug that was uncovered by the new Name module

* stores theories in the knowledge base

A theory is naturally a domain, therefore we can store them in the
knowledge base and create knowledge dependently. This commit also
implements a much faster and simpler version of the join operation,
as well as adds the inheritance notion to the theories.

* updates theory documentation.

* fixes the passes option

It should be possible to specify more than one pass.

* fixes a typo in documentation

which yielded a nasty error message.