fixes the subroutine ordering in the run plugin (#1011)
When `run-entry-points` is set to `all-subroutines`, we launch
all the subroutines in the topological order. The problem is that if a
real callgraph doesn't have a common entry from which all subroutines
are reachable, but instead is represented as a few spanning trees, then
the reverse postorder traverse of a graph will produce an ordering that
is far from the topological. We address this issue by connecting roots
of all trees to an artificial entry point before sorting.

Skip visited

A bonus feature is a new command line option to the run plugin that when
set (on by default) will not revisit a subroutine if it was already
visited. This is only applicable when a set of entry points
(or all entry points) were specified to the run plugin, and it
investigates each in separate.

The justification for this behavior, is that if we already had visited a
subroutine, then we entered it through some other subroutine, i.e., we
investigated it under a richer context, so there is no need to call it
directly. Of course, this is a heuristic,  so it can be disabled:
```
bap /bin/echo --run-entry-points=all-subroutines --run-with-repetitions
```