adds online documentation for knowledge, theories, etc
This PR enhances the `list` command and extends knowledge and
core-theory library interfaces with facilities that let users
inspect classes, properties, agents, and theories, declared in
BAP. This gives more perspective on what is stored in the knowledge
base and how knowledge objects interact.

The list command is extended with four more options:
 - agents
 - theories
 - classes
 - entities

There are also a couple of tweaks to make the list command output
nicer and more intuitve.

A few files were changed mostly because our properties and classes
weren't documented, so I had to go through them and add some
descriptions or refine the existing one.