fixes a plugin initialization issue in baptop, restores postinstall (#1014)
For some yet to be discovered reasons plugins weren't evaluated during
the baptop utility startup. The were loaded, linked, but no
side-effects occured. Not sure what is going in utop internals, but
just loading the "bap.top" library works fine, and it is also
simplifies the implementation.

I've also restored the postinstall hook, that installs manpages for
bap and utilities. The utilities (such as old baptop script and
ppx-bap) are no longer installed, because they are obsolete, but the
man pages are now installed (and now in a proper place).

The postisntall hook is moved to the Makefile, since it has to be run
after plugins are installed. It is predicated with the existence, so
it won't be run on each package installation during opam installation,
only for the bap-std package which actually produces this file.

Right now, I've disabled the installation of the `bap-byteweight`
manpage, since it could be only installed if the corresponding utility
is installed (so we have to create a separate post-install for this package).