adds unknown architecture (#1016)
* adds unknown architecture

This PR introduces an `unknown` architecture to enable analysis of
programs that are not binary, such as JVM or CLR bytecodes. This
change is also in line with the LLVM which also has the unknown
architecture. Additionally, the unkwown architecture could be later
refined, since it is the minimal element of the architecture
domain. That let us develop analyses that will detect architectures,
thus making the architecture a target of analysis rather than a
constant input.

In general, in BAP 2.0 we're moving away from the Arch.t type for
encoding a wide range of architectural properties of computing units
which we model. Mostly because those properties are mostly structural
and could be hardly captured with just a nominal string (even if this
string is triple).

This PR also sweeps through code and removes some warnings as well as
documents the new functions in bap.mli.

* updated testsuite/bap-veri submodule

* fixed warnings