adds --show-invalid, --stop-on-error to mc/objdump, fixes show-size
Now `bap mc` and `bap objdump` will not stop on errors and continue
disassembling unless `--stop-on-error` option is passed. A new option
was added, that will print something when an invalid instruction is
met. If it not specified invalid instructions are silenced.

Fixes options `--show-size` and `--show-kinds` which weren't ouputting
anything due to the incorrect defaults.