makes Knowledge.run more efficeint, bumps version (#1024)
The Knowlege.run was computing every possible slot of a value, to
ensure that all stored computations are triggered. However, for
classes that have a lot of slots it was unnecessary slow, especially
when none of those slots are having any stored computations.

A simple fix, is to ignore slots without stored procedures turns out
int x4 improvement of the run function for the program class. It
doesn't affect the disassemble commmand, which wasn't using the run
command, but affects `mc` and `objdump` which were using it.