implements build caching for plugins
Now all built plugins are cached inside the _build folder, instead of
being rebuilt every time in a temporary folder. And md5sum of the
transitive closure of plugin dependencies is used as the cache
validator. If any dependency is changed, then the plugin is relinked
and repacked.

Additionally, the script builds only those plugins that were migrated
into the build tree, so when we're installing via opam it doesn't try
to rebuild already installed plugins.

This is an extperimental patch, I'm concerned with its
portability. We use md5sum utility, which is not present on MacOS X.
However, there is md5 there, with a slightly different interface,
which we're trying to use instead. Probably, we should rewrite this
in OCaml instead, for better portability and performance.

Finally, it introduces new system dependency, md5[sum], so our docker
builds may fail.